Smaller map added
Files:
1) Graph.m.- I added the originalMap variable and the conditional when
the robot is inside the map and outside the border, when that happens
the robot needs to go to the closest vertex.
2) localise.m.- added the modifiedMap variable with a width size of 5
units, also I comment the line 76 "pose(3)=mod(pose(3), 2*pi); I don't
know what it does but works better without it.
3) MapBorder.m.- fix when calling LineIntersection
4) PathPlan.m.- now passes two maps, the original one and the modified
where the robot will make a route.